feat: perfil and politica migrations, fixed circle dependency, add source-map-support to node

+++ src/server/components/Ticket/ticket.service.ts
@@ -1,15 +1,15 @@
 import { Injectable, BadRequestException, NotFoundException } from '@nestjs/common';
 
 import { InjectModel, InjectSequelize, Sequelize } from '@/server/components/Database';
-import { CreateOrUpdateTicketDto } from './ticket.dto';
+import { TicketInput } from './ticket.dto';
 import { Ticket } from './ticket.entity';
 import { Label } from '@/server/components/Label';
 
 @Injectable()
 export class TicketService {
   public constructor(
-    @InjectModel(Ticket) private readonly ticketsRepository: typeof Ticket,
-    @InjectModel(Label) private readonly labelRepository: typeof Label,
+    @InjectModel(() => Ticket) private readonly ticketsRepository: typeof Ticket,
+    @InjectModel(() => Label) private readonly labelRepository: typeof Label,
     @InjectSequelize() private readonly sequelize: Sequelize
   ) {}
 
@@ -29,12 +29,12 @@ export class TicketService {
     }
   }
 
-  public async createOrUpdate({ label, ...rest }: CreateOrUpdateTicketDto, id?: string) {
+  public async createOrUpdate({ label, ...rest }: TicketInput, id?: string) {
     try {
       if (!id) {
         return this.sequelize.transaction(async t => {
           const ticket = await this.ticketsRepository.create(rest, { transaction: t });
-          if (label.length) {
+          if (label && label.length) {
             await Promise.all(
               label.map(async l => this.labelRepository.create({ ...l, ticketID: ticket.id }, { transaction: t }))
             );
@@ -49,7 +49,7 @@ export class TicketService {
           throw new NotFoundException('Falha ao encontrar ticket');
         }
 
-        if (label.length) {
+        if (label && label.length) {
           await this.labelRepository.destroy({ where: { ticketID: ticket.id }, transaction: t });
           const labels = await Promise.all(
             label.map(async l => this.labelRepository.create({ ...l, ticketID: ticket.id }, { transaction: t }))
